+++ src/flashcardapp/src/App.css
@@ -13,7 +13,7 @@
   box-shadow: 0 0 5px 2px rgba(0, 0, 0, .3);
   box-sizing: border-box; /* Includes padding in width/height */
   background-color: white;
-  width: 150px;
+  width: 200px;
   aspect-ratio: 1 / 1; /* Maintain a 1:1 aspect ratio (square) */
   margin: 10px; /* Space between cards */
   flex-direction: column;
