+++ src/flashcardapp/src/Components/FlashcardList.js
@@ -5,7 +5,7 @@ import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Nav from 'react-bootstrap/Nav';
 
-import Flashcard from './Flashcard';
+import Flashcard from './flashcard/Flashcard';
 import '../App.css';
 
 export default function FlashcardList({ flashcards }) {
