+++ src/flashcardapp/src/Views/carddisplay/carddisplay.modules.css
@@ -1,0 +1,37 @@
+.container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    align-self: center;
+    width: 100%;
+    height: 100%;
+    padding: 100px;
+    align-self: center;
+
+}
+
+
+
+.row {
+    padding: 10px;
+    align-items: center;
+    width: 100%;
+    display: flex;
+    justify-content: space-around;
+}
+
+button {
+    width: 200px;
+    height: 40px;
+    font-size: 16px;
+    align-self: center;
+}
+
+.col {
+    padding: 10px;
+    align-items: center;
+    width: 100%;
+    display: flex;
+    justify-content: space-around;
+}
